github_issue_triage_auto_labeler/requirements.txt
PyGithub>=2.2
regex>=2024.5.15
